+++ lib/src/os_utils.dart
@@ -10,7 +10,8 @@ String expandOS(String os) {
   return os;
 }
 
-String tunDeviceName(String os, int id) {
+String tunDeviceName(String os, String domain) {
+  final id = (domain.hashCode.abs() % 16) + 16;
   if (os == 'darwin') {
     return 'utun$id';
   }